requirements/base.txt
fastapi
uvicorn
sqlalchemy
pydantic

matplotlib
numpy
sympy
scipy

azure-storage-blob